refactor: helpers file
